added compactable interface for aggregate table compaction
