+++ integration/spark2/src/main/scala/org/apache/carbondata/spark/rdd/CompactionFactory.scala
@@ -1,0 +1,53 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *    http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+package org.apache.carbondata.spark.rdd
+
+import java.util.concurrent.ExecutorService
+
+import org.apache.spark.sql.SQLContext
+import org.apache.spark.sql.execution.command.CompactionModel
+
+import org.apache.carbondata.processing.loading.model.CarbonLoadModel
+
+object CompactionFactory {
+
+  /**
+   *  Returns appropriate Compactable object.
+   */
+  def getCompactor(carbonLoadModel: CarbonLoadModel,
+      compactionModel: CompactionModel,
+      executor: ExecutorService,
+      sqlContext: SQLContext,
+      storeLocation: String): Compactor = {
+    if (carbonLoadModel.getCarbonDataLoadSchema.getCarbonTable.isChildDataMap) {
+      new AggregateDataMapCompactor(
+        carbonLoadModel,
+        compactionModel,
+        executor,
+        sqlContext,
+        storeLocation)
+    } else {
+      new CarbonTableCompactor(
+        carbonLoadModel,
+        compactionModel,
+        executor,
+        sqlContext,
+        storeLocation)
+    }
+  }
+}
